- adding new PrefixCache template class   - sync updates to test fwlog app

+++ common/tcanetpp/test/pt_test.cpp
@@ -66,7 +66,7 @@ void debugNodeHandler ( ptNode_t * node )
     return;
 }
 
-void nodeFreeHandler ( uint32_t addr, uint8_t mb, void * rock )
+void nodeFreeHandler ( uint32_t addr, uint16_t mb, void * rock )
 {
     Prefix* p = (Prefix*) rock;
     if ( p ) {
